+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.29.1
+
+### Fixed
+- `Transaction.fromBytes()` and `Transaction.getSignatures()` throw an exception for transactions signed using an ECDSA key
+- spurious `INVALID_TRANSACTION_START` in `TransactionId.generate()`
+
 ## 2.29.0
 
 ### Added